Update Bob the Builder - SoloLearn.ipynb
This Jupyter Notebook contains an exercise for developing a Logistic Regression model. 
Input format:
First line: number of data points in the feature matrix (n)
Next n lines: values of the row in the feature matrix, separated by spaces
Next line: target values separated by spaces
Final line: values of a single datapoint without a target value

The goal is to make a prediction (1 or 0) for the single datapoint.